+++ ICD.Connect.Conferencing.Zoom/Components/Camera/CameraComponent.cs
@@ -1,29 +1,51 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using ICD.Common.Utils;
+using ICD.Common.Properties;
+using ICD.Common.Utils.Collections;
 using ICD.Common.Utils.Extensions;
 using ICD.Connect.Conferencing.Zoom.Responses;
 
 namespace ICD.Connect.Conferencing.Zoom.Components.Camera
 {
-	public class CameraComponent : AbstractZoomRoomComponent
+	public sealed class CameraComponent : AbstractZoomRoomComponent
 	{
 		public event EventHandler OnCamerasUpdated;
 		public event EventHandler OnActiveCameraUpdated;
 
-		private CameraInfo[] m_Cameras;
+		private readonly IcdOrderedDictionary<string, CameraInfo> m_Cameras;
+
 		private string m_SelectedUsbId;
 
-		public CameraInfo ActiveCamera { get { return m_Cameras == null ? null : m_Cameras.SingleOrDefault(c => c.UsbId == m_SelectedUsbId); } }
-
-		public CameraComponent(ZoomRoom parent) : base(parent)
+		/// <summary>
+		/// Gets the active camera.
+		/// </summary>
+		[CanBeNull]
+		public CameraInfo ActiveCamera
 		{
+			get { return m_SelectedUsbId == null ? null : m_Cameras.GetDefault(m_SelectedUsbId); }
+		}
+
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		public CameraComponent(ZoomRoom parent)
+			: base(parent)
+		{
+			m_Cameras = new IcdOrderedDictionary<string, CameraInfo>();
+
 			Subscribe(parent);
 		}
 
+		/// <summary>
+		/// Release resources.
+		/// </summary>
 		protected override void DisposeFinal()
 		{
+			OnCamerasUpdated = null;
+			OnActiveCameraUpdated = null;
+
 			base.DisposeFinal();
 
 			Unsubscribe(Parent);
@@ -39,13 +61,7 @@ namespace ICD.Connect.Conferencing.Zoom.Components.Camera
 
 		public IEnumerable<CameraInfo> GetCameras()
 		{
-			return m_Cameras == null ? Enumerable.Empty<CameraInfo>() : m_Cameras.ToArray(m_Cameras.Length);
-		}
-
-		public void SetNearCameraAsVideoSource(int address)
-		{
-			address = MathUtils.Clamp(address, 0, m_Cameras.Length - 1);
-			SetActiveCameraByUsbId(m_Cameras[address].UsbId);
+			return m_Cameras.Values.ToArray(m_Cameras.Count);
 		}
 
 		public void SetActiveCameraByUsbId(string usbId)
@@ -71,13 +87,27 @@ namespace ICD.Connect.Conferencing.Zoom.Components.Camera
 
 		private void CameraListCallback(ZoomRoom zoomroom, VideoCameraLineResponse response)
 		{
-			m_Cameras = response.Cameras;
+			if (response.Cameras == null)
+				return;
+
+			m_Cameras.Clear();
+			m_Cameras.AddRange(response.Cameras.Select(c => new KeyValuePair<string, CameraInfo>(c.UsbId, c)));
+
 			OnCamerasUpdated.Raise(this);
 		}
 
 		private void SelectedCameraCallback(ZoomRoom zoomRoom, VideoConfigurationResponse response)
 		{
-			m_SelectedUsbId = response.Video.Camera.SelectedId;
+			var video = response.Video;
+			if (video == null)
+				return;
+
+			var camera = video.Camera;
+			if (camera == null)
+				return;
+
+			m_SelectedUsbId = camera.SelectedId;
+
 			OnActiveCameraUpdated.Raise(this);
 		}