+++ ICD.Connect.Conferencing.Zoom/Responses/VideoConfigurationResponse.cs
@@ -1,4 +1,5 @@
-﻿using ICD.Connect.Conferencing.Zoom.Responses.Attributes;
+﻿using ICD.Common.Properties;
+using ICD.Connect.Conferencing.Zoom.Responses.Attributes;
 using ICD.Connect.Conferencing.Zoom.Responses.Converters;
 using Newtonsoft.Json;
 
@@ -9,18 +10,21 @@ namespace ICD.Connect.Conferencing.Zoom.Responses
 	[JsonConverter(typeof(VideoConfigurationResponseConverter))]
 	public sealed class VideoConfigurationResponse : AbstractZoomRoomResponse
 	{
+		[CanBeNull]
 		public VideoConfiguration Video { get; set; }
 	}
 
 	[JsonConverter(typeof(VideoConfigurationConverter))]
 	public sealed class VideoConfiguration
 	{
+		[CanBeNull]
 		public VideoCameraConfiguration Camera { get; set; }
 	}
 
 	[JsonConverter(typeof(VideoCameraConfigurationConverter))]
 	public class VideoCameraConfiguration
 	{
+		[CanBeNull]
 		public string SelectedId { get; set; }
 	}
 }